Infosys-Agentic-Foundry-Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
arize-phoenix==10.6.2
arize-phoenix-otel==0.10.2
astor==0.8.1
asyncpg==0.30.0
attrs==25.4.0
Authlib==1.6.0
azure-ai-documentintelligence==1.0.2
azure-cognitiveservices-speech
bcrypt==4.3.0
black
cachetools==5.5.1
cffi==1.17.1
charset-normalizer==3.4.3
click>=8.1.8
colorama==0.4.6
cryptography==44.0.1
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.10.0
fastapi>=0.110.3
fastmcp==2.3.3
Flask==3.1.0
frozenlist==1.8.0
rapidfuzz==3.14.3
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-adk==1.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.4
googleapis-common-protos==1.71.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.2
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
#kafka
kafka-python==2.3.0

langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.78
langchain-google-genai==2.0.10
langchain-mcp-adapters==0.1.7
langchain-openai==0.3.35
langchain-pymupdf4llm==0.4.1
langchain-text-splitters==0.3.11
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-checkpoint-postgres==2.0.21
langgraph-checkpoint-sqlite==2.0.10
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.33
litellm==1.77.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.13.1
mdurl==0.1.2
motor==3.7.1
mpmath==1.3.0
multidict==6.7.0
mysql-connector-python==9.4.0
mypy-extensions==1.1.0
nltk==3.9.1
numpy>=2.2.4
openai==2.2.0
openinference-instrumentation==0.1.32
openinference-instrumentation-langchain==0.1.43
openpyxl==3.1.5
opentelemetry-api>=1.33.1
opentelemetry-exporter-otlp>=1.33.1
opentelemetry-exporter-otlp-proto-common>=1.33.1
opentelemetry-exporter-otlp-proto-grpc>=1.33.1
opentelemetry-exporter-otlp-proto-http>=1.33.1
opentelemetry-instrumentation>=0.54b1
opentelemetry-proto>=1.33.1
opentelemetry-sdk>=1.33.1
opentelemetry-semantic-conventions>=0.54b1
orjson==3.11.3
ormsgpack==1.10.0
packaging>=24.2
pandas==2.2.3
paramiko
pillow==10.4.0
propcache>=0.2.0
proto-plus==1.25.0
protobuf==5.29.3
psutil==7.0.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.5
pymupdf4llm==0.0.19
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv>=1.0.1
python-multipart==0.0.20
pytz==2024.2
pywin32==310; platform_system == "Windows"
PyYAML==6.0.3
redis
referencing==0.36.2
regex==2025.9.18
requests>=2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.27.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlean.py==3.49.1
starlette>=0.37.2
sympy==1.13.1
tenacity>=8.0.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.5.0
uvicorn==0.34.0
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
azure-storage-blob
boto3
google-cloud-storage
# cloud-detect
# #awscli
# fsspec
# s3fs
# adlfs
